Allow students to be specified by user id
